requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
azure-identity==1.16.0
pyodbc==5.1.0
msal==1.28.0
requests==2.32.0
